+++ test/Journalist.EventStore.IntegrationTests/Streams/EventStreamConsumerTests.cs
@@ -49,43 +49,6 @@ namespace Journalist.EventStore.IntegrationTests.Streams
         }
 
         [Theory, AutoMoqData]
-        public async Task ReceiveEventsAsync_WhenConsumerConfiguredStartReadingFromTheEnd_SkipsPreviousPublishedEvents(
-            string consumerName,
-            JournaledEvent[] dummyEvents)
-        {
-            await PublishEventsAsync(dummyEvents);
-
-            var consumer = await Connection.CreateStreamConsumerAsync(config => config
-                .ReadStream(StreamName, true)
-                .WithName(consumerName));
-
-            Assert.Equal(ReceivingResultCode.EmptyStream, await consumer.ReceiveEventsAsync());
-        }
-
-        [Theory, AutoMoqData]
-        public async Task ReceiveEventsAsync_WhenConsumerConfiguredStartReadingFromTheEnd_ContinuesReadingEndOfStream(
-            string consumerName,
-            JournaledEvent[] dummyEvents)
-        {
-            await PublishEventsAsync(dummyEvents);
-
-            var consumer = await Connection.CreateStreamConsumerAsync(config => config
-                .ReadStream(StreamName, true)
-                .WithName(consumerName));
-
-            Assert.Equal(ReceivingResultCode.EmptyStream, await consumer.ReceiveEventsAsync());
-
-            await PublishEventsAsync(dummyEvents);
-
-            consumer = await Connection.CreateStreamConsumerAsync(config => config
-                .ReadStream(StreamName, true)
-                .WithName(consumerName));
-
-            Assert.Equal(ReceivingResultCode.EventsReceived, await consumer.ReceiveEventsAsync());
-            Assert.Equal(dummyEvents, consumer.EnumerateEvents());
-        }
-
-        [Theory, AutoMoqData]
         public async Task CloseAsync_SavesConsumedPositionPosition(
             JournaledEvent[] dummyEvents1,
             JournaledEvent[] dummyEvents2,